# apps/api/internal/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OPC API",
    "description": "OPC is a content production workspace for one-person video studios. This API exposes CRUD endpoints for the core entities (topics, scripts, content items, knowledge base, series), the IP template generator, AI-assisted tools (topic generation, script humanizer, postmortem), and JSON-based data import/export. The MVP ships without authentication; deploy behind a reverse proxy that gates traffic to trusted clients.\n\nAll requests and responses are JSON. Time fields use RFC 3339. Lists are paginated via `limit` and `offset` query parameters and capped at 200 rows per request.",
    "version": "0.1.0",
    "contact": {
      "name": "OPC",
      "url": "https://github.com/opc/opc"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
...
  "paths": {
...
    "/ai/topics": {
      "post": {
        "tags": ["ai"],
        "summary": "Generate topic ideas with Claude",
        "description": "Asks Claude for `count` topic ideas (capped at 20) seeded from the given `seed` and `platform`. Requires `ANTHROPIC_API_KEY` to be set on the server.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AIGenerateTopicsRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Generated topics",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AIGenerateTopicsResponse"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "502": {"description": "AI returned output that could not be parsed"},
          "503": {"description": "AI service unavailable (no API key, or upstream error)"}
        }
      }
    },
    "/ai/humanize": {
      "post": {
        "tags": ["ai"],
        "summary": "Humanize a script",
        "description": "Rewrites the given script to read less like AI-generated text. Requires `ANTHROPIC_API_KEY`.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AIHumanizeRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Rewritten script",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AIHumanizeResponse"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "503": {"description": "AI service unavailable (no API key, or upstream error)"}
        }
      }
    },
...
    "/ai/postmortem": {
      "post": {
        "tags": ["ai"],
        "summary": "Generate a postmortem for a content item",
        "description": "Walks ContentItem -> Script -> Topic to gather context, then asks Claude for a structured postmortem (success factors, reusable patterns, insights, suggestions). Requires `ANTHROPIC_API_KEY`.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AIPostmortemRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Postmortem report",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AIPostmortemResponse"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"description": "Content item not found"},
          "503": {"description": "AI service unavailable (no API key, or upstream error)"}
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "Invalid request body or query parameter",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {"type": "string", "description": "Human-readable error message"}
        }
      },
...
      "AIGenerateTopicsRequest": {
        "type": "object",
        "required": ["seed", "platform"],
        "properties": {
          "seed": {"type": "string", "description": "Brainstorming seed (e.g. \"夏日复古穿搭\")"},
          "platform": {"type": "string", "description": "Target platform"},
          "count": {"type": "integer", "default": 5, "minimum": 1, "maximum": 20}
        }
      },
      "AIGeneratedTopic": {
        "type": "object",
        "properties": {
          "title": {"type": "string"},
          "angle": {"type": "string"},
          "expected_performance": {"type": "string"},
          "hook": {"type": "string"}
        }
      },
      "AIGenerateTopicsResponse": {
        "type": "object",
        "required": ["topics"],
        "properties": {
          "topics": {"type": "array", "items": {"$ref": "#/components/schemas/AIGeneratedTopic"}}
        }
      },
      "AIHumanizeRequest": {
        "type": "object",
        "required": ["script"],
        "properties": {
          "script": {"type": "string"}
        }
      },
      "AIHumanizeResponse": {
        "type": "object",
        "required": ["humanized"],
        "properties": {
          "humanized": {"type": "string"}
        }
      },
...
      "AIPostmortemRequest": {
        "type": "object",
        "required": ["content_item_id"],
        "properties": {
          "content_item_id": {"type": "integer", "format": "int64"}
        }
      },
      "AIPostmortemStructured": {
        "type": "object",
        "properties": {
          "success_factors": {"type": "array", "items": {"type": "string"}},
          "reusable_patterns": {"type": "array", "items": {"type": "string"}},
          "insights": {"type": "array", "items": {"type": "string"}},
          "suggestions": {"type": "array", "items": {"type": "string"}}
        }
      },
      "AIPostmortemResponse": {
        "type": "object",
        "required": ["report"],
        "properties": {
          "report": {"type": "string", "description": "Raw postmortem text"},
          "structured": {"$ref": "#/components/schemas/AIPostmortemStructured"}
        }
      },
...
    }
  }
}