# apps/api/internal/handlers/ai.go
package handlers

import (
	"context"
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"
	"regexp"
	"strings"
	"time"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"

	"github.com/opc/api/internal/agents"
	"github.com/opc/api/internal/config"
	"github.com/opc/api/internal/models"
)

// demoModeHeader is set on responses served from canned data so the
// frontend (and curl users) can tell the difference between a real
// MiniMax call and a demo response. The header name is fixed across
// all three /ai/* endpoints for grep-ability.
const demoModeHeader = "X-Demo-Mode"

// demoQueryParam is the query-string flag that forces demo mode even
// when an API key is configured. Useful for sales/investor demos
// where the operator wants predictable output regardless of whether
// the key is set.
const demoQueryParam = "demo"

// aiTimeout caps the time a single /ai/* call can spend waiting on
// the MiniMax HTTP API. The default is generous (60s) to allow
// for slow large-context completions; the underlying
// *agents.MiniMax HTTP client also applies its own 10-minute
// per-call timeout, so this is a defense-in-depth upper bound.
const aiTimeout = 60 * time.Second

// maxAIResultBytes caps the raw MiniMax response we are willing to
// scan with a regex when stripping markdown fences. The prompt asks
// for a JSON array of N short topics; even N=50 fits comfortably
// under 1 MiB.
const maxAIResultBytes = 1 << 20

// AIHandler exposes AI-assisted endpoints. The /ai/topics endpoint
// returns N differentiated topic ideas generated by MiniMax for a
// given seed + platform. /ai/humanize rewrites AI-flavored scripts
// to read more naturally. /ai/postmortem walks a content item's FK
// chain to assemble context, then asks MiniMax for a structured
// retrospective.
type AIHandler struct {
	text   *agents.MiniMax
	db     *gorm.DB
	logger *slog.Logger
}

// NewAIHandler wires an AIHandler. A nil logger falls back to
// slog.Default() so callers in main.go do not have to plumb one in.
// The db is optional: it is only required by the /ai/postmortem
// endpoint (which needs to look up ContentItem + Script + Topic).
// Existing callers (tests, MCP wiring) that do not need postmortem
// can pass nil.
func NewAIHandler(text *agents.MiniMax, db ...*gorm.DB) *AIHandler {
	l := slog.Default()
	var d *gorm.DB
	if len(db) > 0 {
		d = db[0]
	}
	return &AIHandler{text: text, db: d, logger: l}
}

// RegisterRoutes attaches the AI endpoints to the router. The route
// table mirrors the legacy Claude-based handler (now MiniMax) exactly so the
// frontend does not have to learn new paths after the provider swap.
func (h *AIHandler) RegisterRoutes(r gin.IRouter) {
	r.POST("/ai/topics", h.GenerateTopics)
	r.POST("/ai/humanize", h.HumanizeScript)
	r.POST("/ai/postmortem", h.Postmortem)
}

// generatedTopic is the wire shape for a single topic. Field names
// use snake_case to match the rest of the API (see models/topic.go).
//
// The upgraded prompt (see prompts.go: GenerateTopicsPrompt) and the
// demo pool (see DemoTopicsJSON) emit `pattern` and `voice_tags`
// alongside the original four fields. The frontend surfaces them as
// design hints, so dropping them at the wire boundary would silently
// break the "clear result display" deliverable. Both fields are
// optional from the parser's perspective — old MiniMax calls that
// return only the four core fields still parse cleanly.
type generatedTopic struct {
	Title               string   `json:"title"`
	Angle               string   `json:"angle"`
	ExpectedPerformance string   `json:"expected_performance"`
	Hook                string   `json:"hook"`
	Pattern             string   `json:"pattern,omitempty"`
	VoiceTags           []string `json:"voice_tags,omitempty"`
}

// generateTopicsRequest is the JSON body for POST /ai/topics.
// Validation is explicit (rather than relying on Gin's binding
// defaults) so we can return precise error messages.
type generateTopicsRequest struct {
	Seed     string `json:"seed"`
	Platform string `json:"platform"`
	Count    int    `json:"count"`
}

type generateTopicsResponse struct {
	Topics []generatedTopic `json:"topics"`
}

// isDemoRequest returns true when the caller asked for demo mode
// via ?demo=true (or 1/yes). We accept a small set of truthy
// spellings so the operator can use whichever form is convenient
// from curl or the browser address bar.
func isDemoRequest(c *gin.Context) bool {
	switch strings.ToLower(c.Query(demoQueryParam)) {
	case "1", "true", "yes", "on":
		return true
	}
	return false
}

// shouldUseDemo returns true when the handler should short-circuit
// to canned data — either because the operator asked for it
// (?demo=true) or because no MiniMax API key is configured. The
// second branch keeps the dev / CI loop useful when MINIMAX_API_KEY
// is absent.
func (h *AIHandler) shouldUseDemo(c *gin.Context) bool {
	return isDemoRequest(c) || !h.text.Available()
}

// markDemoResponse stamps the X-Demo-Mode header on the response so
// the frontend (and curl users) can tell a canned response apart
// from a real MiniMax call.
func markDemoResponse(c *gin.Context) {
	c.Writer.Header().Set(demoModeHeader, "true")
}

// GenerateTopics — POST /ai/topics
//
// Body: {seed, platform, count}
// 200:  {topics: [...]}
// 400:  invalid body (missing seed/platform, count <= 0)
// 502:  MiniMax returned output we could not parse as JSON
// 503:  MiniMax is not configured (no API key) or returned an error
func (h *AIHandler) GenerateTopics(c *gin.Context) {
	var req generateTopicsRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		h.logger.Warn("invalid ai body", "err", err.Error(), "request_id", c.GetString("request_id"))
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid request body"})
		return
	}
	req.Seed = strings.TrimSpace(req.Seed)
	req.Platform = strings.TrimSpace(req.Platform)
	if req.Seed == "" {
		c.JSON(http.StatusBadRequest, gin.H{"error": "seed is required"})
		return
	}
	if req.Platform == "" {
		c.JSON(http.StatusBadRequest, gin.H{"error": "platform is required"})
		return
	}
	if req.Count <= 0 {
		req.Count = 5
	}
	if req.Count > 20 {
		// Hard cap to keep token usage bounded. The prompt does not
		// impose a limit and a runaway "give me 500" would burn budget.
		req.Count = 20
	}

	if h.shouldUseDemo(c) {
		raw, _ := agents.DemoResponse(agents.DemoOpTopics, len(req.Seed))
		topics, err := parseTopics(raw)
		if err != nil {
			h.logger.Error("ai demo parse failed", "err", err.Error(), "request_id", c.GetString("request_id"))
			c.JSON(http.StatusBadGateway, gin.H{
				"error": "AI demo response could not be parsed: " + err.Error(),
			})
			return
		}
		// Demo pool may have grown beyond req.Count (it is a marketing
		// pool, not a per-request fixture). Mirror the real path and
		// return only the first req.Count topics so the wire contract
		// is consistent regardless of which mode the server is in.
		if len(topics) > req.Count {
			topics = topics[:req.Count]
		}
		markDemoResponse(c)
		c.JSON(http.StatusOK, generateTopicsResponse{Topics: topics})
		return
	}

	prompt := agents.GenerateTopicsPrompt(req.Seed, req.Platform, req.Count)

	ctx, cancel := context.WithTimeout(c.Request.Context(), aiTimeout)
	defer cancel()

	res, err := h.text.Text(ctx, prompt, agents.MiniMaxTextOptions{Model: "MiniMax-M2.7-highspeed"})
	if err != nil {
		h.logger.Error("ai text failed", "err", err.Error(), "request_id", c.GetString("request_id"))
		c.JSON(http.StatusServiceUnavailable, gin.H{
			"error": "AI service failed: " + err.Error(),
		})
		return
	}
	// Stamp cost onto the call_log row that the middleware created.
	// MiniMax M2.7-highspeed is the highspeed variant; cost
	// attribution uses the per-1k-token skill row keyed on
	// SkillMiniMaxM27.
	StampClaudeCost(c, config.SkillMiniMaxM27, res.InputTokens, res.OutputTokens)

	topics, err := parseTopics(res.Text)
	if err != nil {
		h.logger.Error("ai parse failed", "err", err.Error(), "request_id", c.GetString("request_id"))
		c.JSON(http.StatusBadGateway, gin.H{
			"error": "AI returned output that could not be parsed: " + err.Error(),
		})
		return
	}

	c.JSON(http.StatusOK, generateTopicsResponse{Topics: topics})
}

// jsonArrayRE matches a top-level JSON array, allowing optional
// leading/trailing markdown fences (```json ... ```) and surrounding
// prose. The first non-greedy match is used; the caller validates
// that it parses.
var jsonArrayRE = regexp.MustCompile(`(?s)\[.*?\]`)

// parseTopics extracts a []generatedTopic from a raw MiniMax
// response. It is lenient about markdown fences and surrounding
// prose because MiniMax often wraps JSON in ```json ... ``` blocks
// despite explicit instructions.
func parseTopics(raw string) ([]generatedTopic, error) {
	if len(raw) > maxAIResultBytes {
		return nil, errors.New("response too large")
	}
	cleaned := strings.TrimSpace(raw)
	// Strip a single ```json ... ``` wrapper if present.
	if strings.HasPrefix(cleaned, "```") {
		if i := strings.Index(cleaned, "\n"); i >= 0 {
			cleaned = cleaned[i+1:]
		}
		if strings.HasSuffix(cleaned, "```") {
			cleaned = cleaned[:len(cleaned)-3]
		}
		cleaned = strings.TrimSpace(cleaned)
	}
	// If the cleaned output still doesn't start with '[', try to
	// locate the first JSON array in the text.
	if !strings.HasPrefix(cleaned, "[") {
		loc := jsonArrayRE.FindStringIndex(cleaned)
		if loc == nil {
			return nil, errors.New("no JSON array found in response")
		}
		cleaned = cleaned[loc[0]:loc[1]]
	}
	var out []generatedTopic
	if err := json.Unmarshal([]byte(cleaned), &out); err != nil {
		return nil, err
	}
	// Drop entries that are missing the required title — MiniMax
	// occasionally returns an empty placeholder object.
	filtered := out[:0]
	for _, t := range out {
		if strings.TrimSpace(t.Title) == "" {
			continue
		}
		filtered = append(filtered, t)
	}
	return filtered, nil
}

type humanizeRequest struct {
	Script string `json:"script"`
}

type humanizeResponse struct {
	Humanized string `json:"humanized"`
}

// HumanizeScript — POST /ai/humanize
//
// Body: {script}
// 200:  {humanized}
// 400:  missing/empty script
// 503:  MiniMax is not configured (no API key) or returned an error
func (h *AIHandler) HumanizeScript(c *gin.Context) {
	var req humanizeRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		h.logger.Warn("invalid humanize body", "err", err.Error(), "request_id", c.GetString("request_id"))
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid request body"})
		return
	}
	req.Script = strings.TrimSpace(req.Script)
	if req.Script == "" {
		c.JSON(http.StatusBadRequest, gin.H{"error": "script is required"})
		return
	}

	if h.shouldUseDemo(c) {
		raw, _ := agents.DemoResponse(agents.DemoOpHumanize, len(req.Script))
		markDemoResponse(c)
		c.JSON(http.StatusOK, humanizeResponse{Humanized: raw})
		return
	}

	prompt := agents.HumanizeScriptPrompt(req.Script)

	ctx, cancel := context.WithTimeout(c.Request.Context(), aiTimeout)
	defer cancel()

	res, err := h.text.Text(ctx, prompt, agents.MiniMaxTextOptions{Model: "MiniMax-M2.7-highspeed"})
	if err != nil {
		h.logger.Error("humanize text failed", "err", err.Error(), "request_id", c.GetString("request_id"))
		c.JSON(http.StatusServiceUnavailable, gin.H{
			"error": "AI service failed: " + err.Error(),
		})
		return
	}
	StampClaudeCost(c, config.SkillMiniMaxM27, res.InputTokens, res.OutputTokens)

	c.JSON(http.StatusOK, humanizeResponse{Humanized: res.Text})
}

type postmortemRequest struct {
	ContentItemID uint `json:"content_item_id"`
}

type postmortemStructured struct {
	SuccessFactors   []string `json:"success_factors"`
	ReusablePatterns []string `json:"reusable_patterns"`
	Insights         []string `json:"insights"`
	Suggestions      []string `json:"suggestions"`
}

type postmortemResponse struct {
	Report     string               `json:"report"`
	Structured postmortemStructured `json:"structured"`
}

// Postmortem — POST /ai/postmortem
//
// Body: {content_item_id}
// 200:  {report, structured}
// 400:  invalid body (missing id, id <= 0)
// 404:  content item not found
// 503:  MiniMax is not configured (no API key) or returned an error
//
// The handler walks ContentItem → Script → Topic to gather context
// for the prompt. Topic lookup failure is non-fatal: the prompt
// degrades gracefully with an empty topic_angle.
func (h *AIHandler) Postmortem(c *gin.Context) {
	var req postmortemRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		h.logger.Warn("invalid postmortem body", "err", err.Error(), "request_id", c.GetString("request_id"))
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid request body"})
		return
	}
	if req.ContentItemID == 0 {
		c.JSON(http.StatusBadRequest, gin.H{"error": "content_item_id is required"})
		return
	}

	if h.shouldUseDemo(c) {
		raw, _ := agents.DemoResponse(agents.DemoOpPostmortem, int(req.ContentItemID))
		structured := parsePostmortemStructured(raw)
		markDemoResponse(c)
		c.JSON(http.StatusOK, postmortemResponse{Report: raw, Structured: structured})
		return
	}

	if h.db == nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "postmortem requires database access"})
		return
	}

	// Walk the FK chain. We do this with three separate First calls
	// (not Preload joins) so a missing Script or Topic surfaces as
	// gorm.ErrRecordNotFound on the offending call rather than
	// silently returning zero values inside a struct. Each lookup
	// scopes by user_id so a caller cannot run a postmortem on
	// another user's content item.
	userID := UserIDFromContext(c)
	var ci models.ContentItem
	ciQuery := h.db.WithContext(withTimeout(c.Request.Context())).Model(&models.ContentItem{})
	if userID != 0 {
		ciQuery = ciQuery.Where("user_id = ?", userID)
	}
	if err := ciQuery.First(&ci, req.ContentItemID).Error; err != nil {
		if errors.Is(err, gorm.ErrRecordNotFound) {
			c.JSON(http.StatusNotFound, gin.H{"error": "content item not found"})
			return
		}
		h.logger.Error("postmortem: content item lookup failed", "err", err.Error(), "id", req.ContentItemID, "request_id", c.GetString("request_id"))
		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to look up content item"})
		return
	}

	var script models.Script
	if err := h.db.WithContext(withTimeout(c.Request.Context())).First(&script, ci.ScriptID).Error; err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
		h.logger.Error("postmortem: script lookup failed", "err", err.Error(), "script_id", ci.ScriptID, "request_id", c.GetString("request_id"))
		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to look up script"})
		return
	}

	topicAngle := ""
	if script.ID != 0 && script.TopicID != 0 {
		var topic models.Topic
		if err := h.db.WithContext(withTimeout(c.Request.Context())).First(&topic, script.TopicID).Error; err == nil {
			topicAngle = topic.Angle
		}
	}

	prompt := agents.PostmortemPrompt(script.Title, script.Content, ci.PerformanceMetrics, topicAngle)

	ctx, cancel := context.WithTimeout(c.Request.Context(), aiTimeout)
	defer cancel()

	res, err := h.text.Text(ctx, prompt, agents.MiniMaxTextOptions{Model: "MiniMax-M2.7-highspeed"})
	if err != nil {
		h.logger.Error("postmortem text failed", "err", err.Error(), "request_id", c.GetString("request_id"))
		c.JSON(http.StatusServiceUnavailable, gin.H{
			"error": "AI service failed: " + err.Error(),
		})
		return
	}
	StampClaudeCost(c, config.SkillMiniMaxM27, res.InputTokens, res.OutputTokens)

	structured := parsePostmortemStructured(res.Text)
	c.JSON(http.StatusOK, postmortemResponse{Report: res.Text, Structured: structured})
}

var jsonObjectRE = regexp.MustCompile(`(?s)\{.*?\}`)

// parsePostmortemStructured extracts the four-field report shape from
// a raw MiniMax response. It is lenient about markdown fences and
// surrounding prose (same approach as parseTopics) and never errors:
// on parse failure it returns the zero-value struct so the frontend
// still has the raw `report` text to display.
func parsePostmortemStructured(raw string) postmortemStructured {
	if len(raw) > maxAIResultBytes {
		return postmortemStructured{}
	}
	cleaned := strings.TrimSpace(raw)
	if strings.HasPrefix(cleaned, "```") {
		if i := strings.Index(cleaned, "\n"); i >= 0 {
			cleaned = cleaned[i+1:]
		}
		if strings.HasSuffix(cleaned, "```") {
			cleaned = cleaned[:len(cleaned)-3]
		}
		cleaned = strings.TrimSpace(cleaned)
	}
	if !strings.HasPrefix(cleaned, "{") {
		loc := jsonObjectRE.FindStringIndex(cleaned)
		if loc == nil {
			return postmortemStructured{}
		}
		cleaned = cleaned[loc[0]:loc[1]]
	}
	var out postmortemStructured
	if err := json.Unmarshal([]byte(cleaned), &out); err != nil {
		return postmortemStructured{}
	}
	return out
}